+++ src/settings/plugins/ifcfg-rh/nms-ifcfg-rh-writer.c
@@ -33,6 +33,7 @@
 #include <stdio.h>
 
 #include "nm-utils/nm-enum-utils.h"
+#include "nm-utils/nm-io-utils.h"
 #include "nm-manager.h"
 #include "nm-setting-connection.h"
 #include "nm-setting-wired.h"
